Fixes #211.