@hameerabbasi 2019-09-29
### Benchmark

```python
with ua.set_backend(Backend):
    with ua.skip_backend(Backend2):
        %timeit mm()
```

`master`: 394 ns ± 6.04 ns
`rich-compare`: 412 ns ± 5.61 ns