@peterbell10 2019-09-29
Is that benchmarking types that don't specialise the comparison operators?