@hameerabbasi 2019-09-29
Yes, it is. The base case. Although I measured a backend with a simple `__eq__` method which uses the `is` operator for comparison, it doesn't seem to produce a measurable difference.